tests only: add check for linkage, deduplicate code
