+++ street/src/test/java/org/opentripplanner/street/linking/IdenticalCoordinateLinkingTest.java
@@ -3,15 +3,11 @@ package org.opentripplanner.street.linking;
 import static com.google.common.truth.Truth.assertThat;
 import static org.opentripplanner.core.model.id.FeedScopedIdFactory.id;
 import static org.opentripplanner.street.model.StreetModelFactory.intersectionVertex;
-import static org.opentripplanner.street.search.TraverseMode.WALK;
 
-import java.util.List;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.opentripplanner.street.model.StreetModelFactory;
-import org.opentripplanner.street.model.edge.StreetTransitStopLink;
 import org.opentripplanner.street.model.vertex.TransitStopVertex;
-import org.opentripplanner.street.search.TraverseModeSet;
 
 /// Tests that linking a vertex with identical coordinates results in the same vertex being re-used
 /// and no split edges/vertices being created.
@@ -35,20 +31,7 @@ class IdenticalCoordinateLinkingTest {
       .withCoordinate(v1.getLat() + offset, v1.getLat() + offset)
       .withId(id("stop"))
       .build();
-    env
-      .linker()
-      .linkVertexPermanently(
-        stopVertex,
-        new TraverseModeSet(WALK),
-        LinkingDirection.BIDIRECTIONAL,
-        ((vertex, streetVertex) -> {
-          var s = (TransitStopVertex) vertex;
-          return List.of(
-            StreetTransitStopLink.createStreetTransitStopLink(s, streetVertex),
-            StreetTransitStopLink.createStreetTransitStopLink(streetVertex, s)
-          );
-        })
-      );
+    env.linkVertexPermanently(stopVertex);
 
     assertThat(env.graph().listStreetEdges()).hasSize(1);
     assertThat(env.graph().summarizeEdges()).containsExactly(
